+++ CHANGELOG.md
@@ -176,7 +176,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - The examples was updated to show that values should be passed as an array,
     even when there is only one value.
   - `Get-TargetResource` was updated to always return an array for parameter
-    `TraceFlags`.
+    `TraceFlags`, `TraceFlagsToInclude`, and `TraceFlagsToInclude`. _The last_
+    _two properties will always return an empty array._
+  - `Set-TargetResource` was updated to handle a single trace flag in the
+    current state.
+  - `Set-TargetResource` was updated to correctly include or exclude a single
+    flag.
 
 ### Fixed